handleAction: folderColor
