+++ kDrive/UI/Controller/Files/MultipleSelectionFloatingPanelViewController.swift
@@ -119,30 +119,9 @@ class MultipleSelectionFloatingPanelViewController: UICollectionViewController {
             }
         case .folderColor:
             group.enter()
-            if driveFileManager.drive.pack == .free {
-                let driveFloatingPanelController = FolderColorFloatingPanelViewController.instantiatePanel()
-                let floatingPanelViewController = driveFloatingPanelController.contentViewController as? FolderColorFloatingPanelViewController
-                floatingPanelViewController?.rightButton.isEnabled = driveFileManager.drive.accountAdmin
-                floatingPanelViewController?.actionHandler = { _ in
-                    driveFloatingPanelController.dismiss(animated: true) {
-                        StorePresenter.showStore(from: self, driveFileManager: self.driveFileManager)
-                    }
-                }
-                present(driveFloatingPanelController, animated: true)
-            } else {
-                let colorSelectionFloatingPanelViewController = ColorSelectionFloatingPanelViewController(files: files, driveFileManager: driveFileManager)
-                let floatingPanelViewController = DriveFloatingPanelController()
-                floatingPanelViewController.isRemovalInteractionEnabled = true
-                floatingPanelViewController.set(contentViewController: colorSelectionFloatingPanelViewController)
-                floatingPanelViewController.track(scrollView: colorSelectionFloatingPanelViewController.collectionView)
-                colorSelectionFloatingPanelViewController.floatingPanelController = floatingPanelViewController
-                colorSelectionFloatingPanelViewController.completionHandler = { isSuccess in
-                    self.success = isSuccess
-                    group.leave()
-                }
-                dismiss(animated: true) {
-                    self.presentingParent?.present(floatingPanelViewController, animated: true)
-                }
+            FileActionsHelper.folderColor(files: files, driveFileManager: driveFileManager, from: self, presentingParent: presentingParent) { isSuccess in
+                self.success = isSuccess
+                group.leave()
             }
         case .download:
             if files.count > Constants.bulkActionThreshold || allItemsSelected || files.contains(where: \.isDirectory) {
