+++ kDrive/Utils/FileActionsHelper.swift
@@ -19,7 +19,7 @@
 import CocoaLumberjackSwift
 import kDriveResources
 import UIKit
-import kDriveCore
+import kDrive
 
 @MainActor
 public class FileActionsHelper {
@@ -146,5 +146,32 @@ public class FileActionsHelper {
         }
 
         return isFavored
+    }
+
+    public static func folderColor(files: [File], driveFileManager: DriveFileManager, from viewController: UIViewController, presentingParent: UIViewController?, completion: (Bool) -> Void) {
+        if driveFileManager.drive.pack == .free {
+            let driveFloatingPanelController = FolderColorFloatingPanelViewController.instantiatePanel()
+            let floatingPanelViewController = driveFloatingPanelController.contentViewController as? FolderColorFloatingPanelViewController
+            floatingPanelViewController?.rightButton.isEnabled = driveFileManager.drive.accountAdmin
+            floatingPanelViewController?.actionHandler = { _ in
+                driveFloatingPanelController.dismiss(animated: true) {
+                    StorePresenter.showStore(from: self, driveFileManager: self.driveFileManager)
+                }
+            }
+            viewController.present(driveFloatingPanelController, animated: true)
+        } else {
+            let colorSelectionFloatingPanelViewController = ColorSelectionFloatingPanelViewController(files: files, driveFileManager: driveFileManager)
+            let floatingPanelViewController = DriveFloatingPanelController()
+            floatingPanelViewController.isRemovalInteractionEnabled = true
+            floatingPanelViewController.set(contentViewController: colorSelectionFloatingPanelViewController)
+            floatingPanelViewController.track(scrollView: colorSelectionFloatingPanelViewController.collectionView)
+            colorSelectionFloatingPanelViewController.floatingPanelController = floatingPanelViewController
+            colorSelectionFloatingPanelViewController.completionHandler = { isSuccess in
+                completion(isSuccess)
+            }
+            viewController.dismiss(animated: true) {
+                presentingParent?.present(floatingPanelViewController, animated: true)
+            }
+        }
     }
 }
